+++ R/np.plot.engine.bandwidth.R
@@ -405,11 +405,9 @@
                 plot.errors.center = plot.errors.center,
                 sub.supplied = sub.supplied,
                 plot.args = persp.args,
-                eligible = .np_plot_variability_single_panel(
-                  plot.par.mfrow = plot.par.mfrow,
-                  continuous = isTRUE(bws$ncon == 2L),
-                  fixed = !isTRUE(rotate)
-                )
+                context = engine.ctx$annotation,
+                continuous = isTRUE(bws$ncon == 2L),
+                fixed = !isTRUE(rotate)
               )
             )
             if (plot.errors.type == "all" && !is.null(lerr.all) && !is.null(herr.all)) {
@@ -550,6 +548,7 @@
           }
         } else if (plot.behavior != "data") {
           plot.layout <- .np_plot_layout_activate(plot.layout)
+          .np_plot_variability_panel_begin(engine.ctx$annotation)
           ## plot evaluation
           plot.fun <- if (xi.factor) {
             .np_plot_panel_fun(plot.bootstrap = plot.bootstrap, plot.bxp = plot.bxp)
@@ -608,10 +607,8 @@
               plot.errors.center = plot.errors.center,
               sub.supplied = sub.supplied,
               plot.args = plot.args,
-              eligible = .np_plot_variability_single_panel(
-                plot.par.mfrow = plot.par.mfrow,
-                continuous = !xi.factor
-              )
+              context = engine.ctx$annotation,
+              continuous = !xi.factor
             )
             if (plot.errors.type == "all") {
               draw.all.error.types(
@@ -711,6 +708,7 @@
         for (i in seq_len(bws$ndim)){
           xi.factor = is.factor(xdat[,i])
           plot.layout <- .np_plot_layout_activate(plot.layout)
+          .np_plot_variability_panel_begin(engine.ctx$annotation)
 
           ## plot evaluation
           plot.fun <- if (xi.factor) {
@@ -765,10 +763,8 @@
               plot.errors.center = plot.errors.center,
               sub.supplied = sub.supplied,
               plot.args = plot.args,
-              eligible = .np_plot_variability_single_panel(
-                plot.par.mfrow = plot.par.mfrow,
-                continuous = !xi.factor
-              )
+              context = engine.ctx$annotation,
+              continuous = !xi.factor
             )
             if (plot.errors.type == "all") {
               draw.all.error.types(