+++ R/np.plot.engine.conbandwidth.R
@@ -578,15 +578,13 @@
                 plot.errors.center = plot.errors.center,
                 sub.supplied = sub.supplied,
                 plot.args = persp.args,
-                eligible = .np_plot_variability_single_panel(
-                  plot.par.mfrow = plot.par.mfrow,
-                  continuous = if (quantreg) {
-                    isTRUE(bws$xncon == 2L)
-                  } else {
-                    isTRUE(bws$xncon + bws$yncon == 2L)
-                  },
-                  fixed = !isTRUE(rotate)
-                )
+                context = engine.ctx$annotation,
+                continuous = if (quantreg) {
+                  isTRUE(bws$xncon == 2L)
+                } else {
+                  isTRUE(bws$xncon + bws$yncon == 2L)
+                },
+                fixed = !isTRUE(rotate)
               )
             )
             if (plot.errors.type == "all" && !is.null(lerr.all) && !is.null(herr.all)) {
@@ -828,6 +826,7 @@
             }
           } else if (plot.behavior != "data") {
             plot.layout <- .np_plot_layout_activate(plot.layout)
+            .np_plot_variability_panel_begin(engine.ctx$annotation)
             ## plot evaluation
             plot.fun <- if (xi.factor) {
               .np_plot_panel_fun(plot.bootstrap = plot.bootstrap, plot.bxp = plot.bxp)
@@ -893,10 +892,8 @@
                 plot.errors.center = plot.errors.center,
                 sub.supplied = sub.supplied,
                 plot.args = plot.args,
-                eligible = .np_plot_variability_single_panel(
-                  plot.par.mfrow = plot.par.mfrow,
-                  continuous = !xi.factor
-                )
+                context = engine.ctx$annotation,
+                continuous = !xi.factor
               )
               if (plot.errors.type == "all") {
                 draw.all.error.types(
@@ -1112,6 +1109,7 @@
               }
             } else if (plot.behavior != "data") {
               plot.layout <- .np_plot_layout_activate(plot.layout)
+              .np_plot_variability_panel_begin(engine.ctx$annotation)
               ## plot evaluation
               plot.fun <- if (xi.factor) {
                 .np_plot_panel_fun(plot.bootstrap = plot.bootstrap, plot.bxp = plot.bxp)
@@ -1177,10 +1175,8 @@
                   plot.errors.center = plot.errors.center,
                   sub.supplied = sub.supplied,
                   plot.args = plot.args,
-                  eligible = .np_plot_variability_single_panel(
-                    plot.par.mfrow = plot.par.mfrow,
-                    continuous = !xi.factor
-                  )
+                  context = engine.ctx$annotation,
+                  continuous = !xi.factor
                 )
                 if (plot.errors.type == "all") {
                   draw.all.error.types(
@@ -1311,6 +1307,7 @@
 
           for (j in seq_len(dsf)){
             plot.layout <- .np_plot_layout_activate(plot.layout)
+            .np_plot_variability_panel_begin(engine.ctx$annotation)
             ## plot evaluation
             idx <- (plot.index-1)*dsf+j
             plot.fun <- if (xi.factor) {
@@ -1363,10 +1360,8 @@
                 plot.errors.center = plot.errors.center,
                 sub.supplied = sub.supplied,
                 plot.args = plot.args,
-                eligible = .np_plot_variability_single_panel(
-                  plot.par.mfrow = plot.par.mfrow,
-                  continuous = !xi.factor
-                )
+                context = engine.ctx$annotation,
+                continuous = !xi.factor
               )
               if (plot.errors.type == "all") {
                 draw.all.error.types(