+++ R/np.plot.engine.condbandwidth.R
@@ -595,15 +595,13 @@
                 plot.errors.center = plot.errors.center,
                 sub.supplied = sub.supplied,
                 plot.args = persp.args,
-                eligible = .np_plot_variability_single_panel(
-                  plot.par.mfrow = plot.par.mfrow,
-                  continuous = if (quantreg) {
-                    isTRUE(bws$xncon == 2L)
-                  } else {
-                    isTRUE(bws$xncon + bws$yncon == 2L)
-                  },
-                  fixed = !isTRUE(rotate)
-                )
+                context = engine.ctx$annotation,
+                continuous = if (quantreg) {
+                  isTRUE(bws$xncon == 2L)
+                } else {
+                  isTRUE(bws$xncon + bws$yncon == 2L)
+                },
+                fixed = !isTRUE(rotate)
               )
             )
             if (plot.errors.type == "all" && !is.null(lerr.all) && !is.null(herr.all)) {
@@ -787,7 +785,13 @@
           cex.axis = cex.axis,
           cex.lab = cex.lab,
           cex.main = cex.main,
-          cex.sub = cex.sub
+          cex.sub = cex.sub,
+          annotation = .np_plot_variability_annotation_spec(
+            plot.errors.method, plot.errors.type, plot.errors.alpha,
+            plot.errors.center, sub.supplied, plot.args = list(),
+            context = engine.ctx$annotation, continuous = !xi.factor,
+            per.quantile = TRUE
+          )
         )
       }
 
@@ -998,6 +1002,7 @@
             plot.out[[plot.index]]$bxp <- temp.boot
           } else if (plot.behavior != "data") {
             plot.layout <- .np_plot_layout_activate(plot.layout)
+            .np_plot_variability_panel_begin(engine.ctx$annotation)
             if (multi.eval) {
               multi.err <- if (plot.errors) temp.err.arr[seq_len(xi.neval), , , drop = FALSE] else NULL
               plot_multi_tau(
@@ -1119,10 +1124,8 @@
                 plot.errors.center = plot.errors.center,
                 sub.supplied = sub.supplied,
                 plot.args = plot.args,
-                eligible = .np_plot_variability_single_panel(
-                  plot.par.mfrow = plot.par.mfrow,
-                  continuous = !xi.factor
-                )
+                context = engine.ctx$annotation,
+                continuous = !xi.factor
               )
               if (plot.errors.type == "all") {
                 draw.all.error.types(
@@ -1346,6 +1349,7 @@
               }
             } else if (plot.behavior != "data") {
               plot.layout <- .np_plot_layout_activate(plot.layout)
+              .np_plot_variability_panel_begin(engine.ctx$annotation)
               ## plot evaluation
               plot.fun <- if (xi.factor) {
                 .np_plot_panel_fun(plot.bootstrap = plot.bootstrap, plot.bxp = plot.bxp)
@@ -1411,10 +1415,8 @@
                   plot.errors.center = plot.errors.center,
                   sub.supplied = sub.supplied,
                   plot.args = plot.args,
-                  eligible = .np_plot_variability_single_panel(
-                    plot.par.mfrow = plot.par.mfrow,
-                    continuous = !xi.factor
-                  )
+                  context = engine.ctx$annotation,
+                  continuous = !xi.factor
                 )
                 if (plot.errors.type == "all") {
                   draw.all.error.types(
@@ -1550,6 +1552,7 @@
           for (j in seq_len(dsf)){
             grad.j <- if (plot.index <= bws$xndim) gradient_component_index(i, j) else j
             plot.layout <- .np_plot_layout_activate(plot.layout)
+            .np_plot_variability_panel_begin(engine.ctx$annotation)
             ## plot evaluation
             idx <- (plot.index-1)*dsf+j
             plot.fun <- if (xi.factor) {
@@ -1602,10 +1605,8 @@
                 plot.errors.center = plot.errors.center,
                 sub.supplied = sub.supplied,
                 plot.args = plot.args,
-                eligible = .np_plot_variability_single_panel(
-                  plot.par.mfrow = plot.par.mfrow,
-                  continuous = !xi.factor
-                )
+                context = engine.ctx$annotation,
+                continuous = !xi.factor
               )
               if (plot.errors.type == "all") {
                 draw.all.error.types(