+++ R/np.plot.engine.rbandwidth.R
@@ -549,11 +549,9 @@
                 plot.errors.center = plot.errors.center,
                 sub.supplied = sub.supplied,
                 plot.args = persp.args,
-                eligible = .np_plot_variability_single_panel(
-                  plot.par.mfrow = plot.par.mfrow,
-                  continuous = isTRUE(bws$ncon == 2L),
-                  fixed = !isTRUE(rotate)
-                )
+                context = engine.ctx$annotation,
+                continuous = isTRUE(bws$ncon == 2L),
+                fixed = !isTRUE(rotate)
               )
             )
 
@@ -714,6 +712,7 @@
           }
         } else if (plot.behavior != "data") {
           plot.layout <- .np_plot_layout_activate(plot.layout)
+          .np_plot_variability_panel_begin(engine.ctx$annotation)
           ## plot evaluation
           plot.fun <- if (xi.factor) {
             .np_plot_panel_fun(plot.bootstrap = plot.bootstrap, plot.bxp = plot.bxp)
@@ -872,10 +871,8 @@
               plot.errors.center = plot.errors.center,
               sub.supplied = sub.supplied,
               plot.args = plot.args,
-              eligible = .np_plot_variability_single_panel(
-                plot.par.mfrow = plot.par.mfrow,
-                continuous = !xi.factor
-              )
+              context = engine.ctx$annotation,
+              continuous = !xi.factor
             )
             drew.bias.center <- .np_plot_draw_bias_center_1d(
               x = ei,
@@ -1017,6 +1014,7 @@
         for (i in seq_len(bws$ndim)){
           xi.factor = is.factor(xdat[,i])
           plot.layout <- .np_plot_layout_activate(plot.layout)
+          .np_plot_variability_panel_begin(engine.ctx$annotation)
 
           ## plot evaluation
           plot.fun <- if (xi.factor) {
@@ -1157,10 +1155,8 @@
               plot.errors.center = plot.errors.center,
               sub.supplied = sub.supplied,
               plot.args = plot.args,
-              eligible = .np_plot_variability_single_panel(
-                plot.par.mfrow = plot.par.mfrow,
-                continuous = !xi.factor
-              )
+              context = engine.ctx$annotation,
+              continuous = !xi.factor
             )
             drew.bias.center <- .np_plot_draw_bias_center_1d(
               x = allei[,i],