+++ R/np.plot.engine.sibandwidth.R
@@ -315,6 +315,7 @@
 
       if (plot.behavior != "data"){      
         plot.layout <- .np_plot_layout_activate(plot.layout)
+        .np_plot_variability_panel_begin(engine.ctx$annotation)
         if (plot.errors){
           plot.args <- list(x = tobj$index[i.sort],
                             y = temp.mean[i.sort],
@@ -348,9 +349,7 @@
             plot.errors.center = plot.errors.center,
             sub.supplied = sub.supplied,
             plot.args = plot.args,
-            eligible = .np_plot_variability_single_panel(
-              plot.par.mfrow = plot.par.mfrow
-            )
+            context = engine.ctx$annotation
           )
           if (plot.errors.type == "all") {
             sorted.all.err <- lapply(temp.all.err, function(err) {
@@ -476,6 +475,7 @@
         for (i in seq_len(ncol(xdat))) {
           if (plot.behavior != "data"){
             plot.layout <- .np_plot_layout_activate(plot.layout)
+            .np_plot_variability_panel_begin(engine.ctx$annotation)
 
             if (is.null(ylim)) {
               if (!common.scale) {
@@ -524,9 +524,7 @@
                 plot.errors.center = plot.errors.center,
                 sub.supplied = sub.supplied,
                 plot.args = plot.args,
-                eligible = .np_plot_variability_single_panel(
-                  plot.par.mfrow = plot.par.mfrow
-                )
+                context = engine.ctx$annotation
               )
               if (plot.errors.type == "all") {
                 scaled.all.err <- lapply(temp.all.err, function(err) {